Refractor Message type
Make Message an instance of Binary. This will help for serializing and
deserializing in conduit.

We move the `time` field from MessageContext to MessageBody. This is
because it is only really required for some messages, and it is always
shown in the message body. Before we were just parsing the time field
and throwing it out.