Move `lastBlock` to mutable context
Currently, we update `lastBlock` in the interpreter whenever we add ro
delete blocks. This was an improvement over doing it everywhere in the
code, but it still results in duplicated code in our multiple
interpreters which is more error prone.

To avoid this, we can update `lastBlock` in our helper functions in
the Connection' monad. Putting `lastBlock` allows for this change.