Add table of UTXOs to the Overview tab.
The Scotty server now accepts GET requests at '/utxos' and replies
with a JSON list of the data in the persistent_u_t_x_o table. The
Overview tab displays this table. Still need to change this table
to automatically update and ideally also add info about which
utxos are actually in blocks (with # of confirmations).