automatically send new utxo data when needed.
Instead of having the "Update UTXOs" button in the ui, we simply
request MerkleBlock data whenever a new header is detected. Then
when a new block arrives we send a NewBlock message to the ui and
when the utxos are updated we send a UTXOsUpdated message to the
ui. These prompt the ui to update the overview page by sending
a new HTTP GET request to '/utxo'.